refactor: Decompose monolithic useEventData hook into modular architecture
BREAKING CHANGE: Internal architecture completely refactored for maintainability

* Hook Decomposition (1000+ → 516 lines, 48% reduction):
  - useErrorTracking.ts (113 lines) - Centralized error management
  - usePeopleTransform.ts (127 lines) - People data transformation
  - useEventsTransform.ts (158 lines) - Events data transformation
  - useTeamCapacities.ts (133 lines) - Team capacity management
  - useDayCellData.ts (184 lines) - Day cell data with validation

* New Utilities:
  - dataQueries.ts (115 lines) - Pure data query functions

* Benefits:
  - Single responsibility principle applied throughout
  - Enhanced testability with isolated modules
  - Better performance through optimized memoization
  - Improved reusability of error tracking and utilities
  - Clear dependency structure

* Documentation:
  - Updated CHANGELOG.md with v1.12.0 architecture improvements
  - Enhanced CLAUDE.md with modular hook structure
  - Comprehensive JSDoc comments for all new modules

* Backward Compatibility:
  - All public APIs remain unchanged
  - No breaking changes for existing integrations
  - Internal refactoring only
