+++ CLAUDE.md
@@ -31,6 +31,7 @@ This repository contains a **Shift Scheduler widget** development project for Me
 - **Data Model**: People (Engineers) and Event entities with direct Mendix integration  
 - **UI Pattern**: Horizontal scrollable timeline with team-grouped person rows
 - **Interactions**: Double-click editing, context menus, drag-to-scroll navigation
+- **Hook Architecture**: Modular, composable hooks (v1.12.0+) for better maintainability
 
 ### Technology Stack
 ```json
@@ -61,13 +62,24 @@ This repository contains a **Shift Scheduler widget** development project for Me
 - `src/components/LoadingStates.tsx` - Loading, error, and empty state components
 - `src/components/DebugPanel.tsx` - Development debug information panel
 
-### Custom Hooks
-- `src/hooks/useEventData.ts` - Event data management and validation
+### Custom Hooks (Modular Architecture v1.12.0+)
+**Core Data Management:**
+- `src/hooks/useEventData.ts` - Main orchestrator hook (516 lines)
+- `src/hooks/useErrorTracking.ts` - Centralized error management (113 lines) 
+- `src/hooks/usePeopleTransform.ts` - People data transformation (127 lines)
+- `src/hooks/useEventsTransform.ts` - Events data transformation (158 lines)
+- `src/hooks/useTeamCapacities.ts` - Team capacity management (133 lines)
+- `src/hooks/useDayCellData.ts` - Day cell data with validation (184 lines)
+
+**UI Interaction:**
 - `src/hooks/useScrollNavigation.ts` - Horizontal scroll and infinite loading
 - `src/hooks/useMultiSelect.ts` - Multi-selection with keyboard modifiers
 - `src/hooks/useKeyboardNavigation.ts` - Arrow keys and keyboard shortcuts
 - `src/hooks/useContextMenu.ts` - Context menu state and option generation
 - `src/hooks/useTeamGrouping.ts` - Team/lane structure processing
+
+**Utilities:**
+- `src/utils/dataQueries.ts` - Pure data query functions (115 lines)
 
 ### Type Definitions
 - `src/types/shiftScheduler.ts` - TypeScript interfaces for People and Event assignments
